+++ src/routes/(v2)/v2/valibot-nullable/schema.ts
@@ -1,0 +1,6 @@
+import { object, string, minLength, email, nullable } from 'valibot';
+
+export const schema = object({
+	name: nullable(string([minLength(1)])),
+	email: string([email()])
+});