+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.4.0] - 2024-02-20
 
 ### Added
 
@@ -14,8 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Nullable fields weren't always detected in JSON Schemas.
-- Constraints weren't added when using default values for an adapter with introspection.
-- Fixed null support for Joi schemas.
+- Constraints weren't added when using default values for Arktype, Vine and Valibot schemas.
+- Fixed null support for Joi, Arktype and Vine schemas.
 
 ## [2.3.0] - 2024-02-18